+++ chalkbox/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.1.0"
+__version__ = "2.1.1"
 
 from .components.alert import Alert
 from .components.align import Align